# services/kurikulum-nilai/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service C - Prasyarat dan Kurikulum API",
        "description": "API untuk validasi prasyarat dan kurikulum mahasiswa. Service ini menyediakan akses data nilai mahasiswa, IPS semester, dan detail kurikulum.",
        "contact": {
            "name": "Andi Muh. Arif Darma Saputra M",
            "email": "102022580023@student.telkomuniversity.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/kurikulum": {
            "get": {
                "tags": [
                    "Kurikulum"
                ],
                "summary": "Lihat daftar semua kurikulum",
                "description": "Mengambil daftar seluruh mata kuliah dalam kurikulum. Endpoint ini berfungsi sebagai Collection.",
                "operationId": "getKurikulumList",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar kurikulum",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data kurikulum berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "kode_matkul": {
                                                        "type": "string",
                                                        "example": "SI101"
                                                    },
                                                    "nama_matkul": {
                                                        "type": "string",
                                                        "example": "Algoritma dan Pemrograman"
                                                    },
                                                    "sks": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "semester": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "prodi": {
                                                        "type": "string",
                                                        "example": "S1 Sistem Informasi"
                                                    },
                                                    "prasyarat": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "deskripsi": {
                                                        "type": "string",
                                                        "example": "Mata kuliah dasar pemrograman",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Prasyarat-Kurikulum-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - API Key tidak valid"
                    }
                },
                "security": [
                    {
                        "X-IAE-KEY": []
                    }
                ]
            }
        },
        "/api/v1/kurikulum/{kode}": {
            "get": {
                "tags": [
                    "Kurikulum"
                ],
                "summary": "Lihat detail kurikulum berdasarkan kode matkul",
                "description": "Mengambil detail mata kuliah spesifik berdasarkan kode matkul. Dipanggil oleh Service B untuk validasi matkul.",
                "operationId": "getKurikulumByKode",
                "parameters": [
                    {
                        "name": "kode",
                        "in": "path",
                        "description": "Kode mata kuliah",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "SI101"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail kurikulum",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail kurikulum berhasil diambil"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "kode_matkul": {
                                                    "type": "string",
                                                    "example": "SI101"
                                                },
                                                "nama_matkul": {
                                                    "type": "string",
                                                    "example": "Algoritma dan Pemrograman"
                                                },
                                                "sks": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "semester": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "prodi": {
                                                    "type": "string",
                                                    "example": "S1 Sistem Informasi"
                                                },
                                                "prasyarat": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "deskripsi": {
                                                    "type": "string",
                                                    "example": "Mata kuliah dasar pemrograman",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kurikulum tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthorized - API Key tidak valid"
                    }
                },
                "security": [
                    {
                        "X-IAE-KEY": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "X-IAE-KEY": {
                "type": "apiKey",
                "description": "Masukkan NIM sebagai API Key (102022580023)",
                "name": "X-IAE-KEY",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Bearer JWT dari IAE SSO (https://iae-sso.virtualfri.id/api/v1/auth/token)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}